8332252: Clean up vmTestbase/vm/share
