+++ test/hotspot/jtreg/vmTestbase/gc/gctests/LoadUnloadGC/LoadUnloadGC.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2002, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2002, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -71,12 +71,8 @@
 
 package gc.gctests.LoadUnloadGC;
 
-import nsk.share.test.*;
 import nsk.share.gc.*;
 import nsk.share.classload.ClassPathNonDelegatingClassLoader;
-import vm.share.monitoring.MemoryPoolFinder;
-
-import java.io.*;
 import java.util.*;
 import java.lang.management.MemoryPoolMXBean;
