+++ test/hotspot/jtreg/vmTestbase/gc/gctests/LoadUnloadGC/MemoryPoolFinder.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013, 2018, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2013, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -20,11 +20,11 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package vm.share.monitoring;
+package gc.gctests.LoadUnloadGC;
 
 import java.lang.management.*;
 
-public enum MemoryPoolFinder {
+enum MemoryPoolFinder {
     CODE_CACHE,
         EDEN_SPACE,
         SURVIVOR_SPACE,
