+++ test/hotspot/jtreg/vmTestbase/metaspace/share/TriggerUnloadingWithWhiteBox.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011, 2018, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2013, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -20,10 +20,19 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package vm.share.process;
+package metaspace.share;
 
-public interface MessageOutput {
-        public void start();
-        public void send(String msg);
-        public void finish();
+
+import jdk.test.whitebox.WhiteBox;
+import nsk.share.test.ExecutionController;
+
+public class TriggerUnloadingWithWhiteBox implements TriggerUnloadingHelper {
+
+        private final static WhiteBox wb = WhiteBox.getWhiteBox();
+
+        @Override
+        public void triggerUnloading(ExecutionController stresser) {
+                wb.fullGC();
+        }
+
 }
