+++ test/hotspot/jtreg/vmTestbase/metaspace/staticReferences/StaticReferences.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2013, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -59,8 +59,8 @@ import nsk.share.test.ExecutionController;
 import nsk.share.test.Stresser;
 import nsk.share.test.TestBase;
 import nsk.share.test.Tests;
-import vm.share.gc.TriggerUnloadingHelper;
-import vm.share.gc.TriggerUnloadingWithWhiteBox;
+import metaspace.share.TriggerUnloadingHelper;
+import metaspace.share.TriggerUnloadingWithWhiteBox;
 
 /**
  * Test checks that static fields will be initialized in new loaded class. Test performs in loop the following routine:
