+++ test/hotspot/jtreg/vmTestbase/metaspace/stressHierarchy/common/PerformChecksHelper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2013, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -34,7 +34,7 @@ import metaspace.stressHierarchy.common.exceptions.ClassNotUnloadedException;
 import metaspace.stressHierarchy.common.exceptions.TimeIsOverException;
 import nsk.share.test.ExecutionController;
 import jdk.test.whitebox.WhiteBox;
-import vm.share.gc.TriggerUnloadingHelper;
+import metaspace.share.TriggerUnloadingHelper;
 
 public class PerformChecksHelper {
