+++ test/hotspot/jtreg/vmTestbase/metaspace/stressHierarchy/common/StressHierarchyBaseClass.java
@@ -24,10 +24,10 @@ package metaspace.stressHierarchy.common;
 
 import java.net.MalformedURLException;
 
-import vm.share.gc.HeapOOMEException;
-import vm.share.gc.TriggerUnloadingByFillingMetaspace;
-import vm.share.gc.TriggerUnloadingHelper;
-import vm.share.gc.TriggerUnloadingWithWhiteBox;
+import metaspace.share.HeapOOMEException;
+import metaspace.share.TriggerUnloadingByFillingMetaspace;
+import metaspace.share.TriggerUnloadingHelper;
+import metaspace.share.TriggerUnloadingWithWhiteBox;
 
 import metaspace.stressHierarchy.common.classloader.tree.Node;
 import metaspace.stressHierarchy.common.classloader.tree.Tree;
