+++ test/hotspot/jtreg/vmTestbase/vm/compiler/complog/share/LogCompilationTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2013, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -29,7 +29,6 @@ import nsk.share.log.Log;
 import nsk.share.log.LogSupport;
 import vm.share.options.Option;
 import vm.share.options.OptionSupport;
-import vm.share.process.ProcessExecutor;
 
 import java.io.File;
 import java.io.FileNotFoundException;
