+++ test/hotspot/jtreg/vmTestbase/vm/compiler/complog/share/StreamListener.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2011, 2018, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -20,17 +20,17 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package vm.share.gc;
+package vm.compiler.complog.share;
 
-/**
- * This class is used to distinguish between OOME in metaspace and OOME in heap when triggering class unloading.
+/*
+ * StreamListener listens on events from BufferedInputStream.
+ *
+ * Note: StreamListener should not never block as it potentially
+ * runs in thread that reads the input.
  */
-public class HeapOOMEException extends RuntimeException {
-
-    private static final long serialVersionUID = 1L;
-
-    public HeapOOMEException(String string) {
-        super(string);
-    }
-
+public interface StreamListener {
+        public void onStart();
+        public void onRead(String line);
+        public void onFinish();
+        public void onException(Throwable e);
 }
