+++ test/hotspot/jtreg/vmTestbase/vm/mlvm/hiddenloader/stress/byteMutation/Test.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2010, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -44,7 +44,7 @@ package vm.mlvm.hiddenloader.stress.byteMutation;
 
 import vm.mlvm.hiddenloader.share.HiddenkTestee01;
 import vm.mlvm.hiddenloader.share.StressClassLoadingTest;
-import vm.share.FileUtils;
+import vm.mlvm.share.FileUtils;
 import vm.share.options.Option;
 
 /**
