+++ test/hotspot/jtreg/vmTestbase/vm/mlvm/share/CustomClassLoaders.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011, 2018, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2011, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -24,8 +24,6 @@
 package vm.mlvm.share;
 
 import java.io.IOException;
-
-import vm.share.FileUtils;
 
 public class CustomClassLoaders {
