--- test/hotspot/jtreg/vmTestbase/vm/share/transform/TransformingClassLoader.java
@@ -1,61 +1,0 @@
-/*
- * Copyright (c) 2013, 2018, Oracle and/or its affiliates. All rights reserved.
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
- *
- * This code is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License version 2 only, as
- * published by the Free Software Foundation.
- *
- * This code is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
- * version 2 for more details (a copy is included in the LICENSE file that
- * accompanied this code).
- *
- * You should have received a copy of the GNU General Public License version
- * 2 along with this work; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- *
- * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
- * or visit www.oracle.com if you need additional information or have any
- * questions.
- */
-package vm.share.transform;
-
-import vm.share.FileUtils;
-
-public class TransformingClassLoader extends ClassLoader {
-    private final AbstractClassFileTransformer transformer;
-
-    protected TransformingClassLoader(ClassLoader parent,
-            AbstractClassFileTransformer transformer) {
-        super(parent);
-        this.transformer = transformer;
-    }
-
-    @Override
-    protected Class<?> loadClass(String name, boolean resolve)
-            throws ClassNotFoundException {
-        if (!transformer.shouldBeTransformed(name)) {
-            return super.loadClass(name, resolve);
-        }
-        synchronized (getClassLoadingLock(name)) {
-            // First, check if the class has already been loaded
-            Class<?> c = findLoadedClass(name);
-            if (c == null) {
-                try {
-                    byte[] bytes = FileUtils.readClass(name);
-                    bytes = transformer.transformClass(bytes);
-                    c = defineClass(name, bytes, 0, bytes.length);
-                } catch (Exception e) {
-                    e.printStackTrace();
-                    return super.loadClass(name, resolve);
-                }
-            }
-            if (resolve) {
-                resolveClass(c);
-            }
-            return c;
-        }
-    }
-}
